Added state files properly
