--- Rating.cpp
@@ -1,16 +1,0 @@
-/**
- * @file Rating.cpp
- * @brief Contains implementation for the Rating class
- * @authors Sange Tshakumane (u21479748)
-*/
-
-#include <iostream>
-#include "Rating.h"
-using namespace std;
-
-/**
- * Destructor function for the rating class
-*/
-Rating::~Rating() {
-    cout << "Rating destructor called" << endl;
-}
