+++ code/src/Neutral.h
@@ -1,10 +1,12 @@
 /**
  * @file Neutral.h
  * @brief Contains declaration for the Neutral class
- * @authors Aidan Chapman (u22738917)
+ * @authors Aidan Chapman (u22738917), Sange Tshakumane (u21479748)
 */
 
 #include "Rating.h"
+
+using namespace std;
 
 #ifndef NEUTRAL_H
 #define NEUTRAL_H
@@ -15,6 +17,8 @@ class Neutral : public Rating
         Neutral();
         ~Neutral();
         float calculateTip();
+        void changeState(Customer* customer);
+        string getRating();
 };
 
 #endif
