+++ code/src/Rating.h
@@ -1,12 +1,16 @@
 /**
  * @file Rating.h
  * @brief Contains declaration for the Rating class
- * @authors Aidan Chapman (u22738917)
+ * @authors Aidan Chapman (u22738917), Sange Tshakumane (u21479748)
 */
+
+#include "Customer.h"
+
+using namespace std;
 
 #ifndef RATING_H
 #define RATING_H
-class Customer;
+
 class Rating{
     public:
         Rating();
