+++ code/src/Satisfied.h
@@ -1,10 +1,12 @@
 /**
  * @file Satisfied.h
  * @brief Contains declaration for the Satisfied class
- * @authors Aidan Chapman (u22738917)
+ * @authors Aidan Chapman (u22738917), Sange Tshakumane (u21479748)
 */
 
 #include "Rating.h"
+
+using namespace std;
 
 #ifndef SATISFIED_H
 #define SATISFIED_H
@@ -15,6 +17,8 @@ class Satisfied : public Rating
         Satisfied();
         ~Satisfied();
         float calculateTip();
+        void changeState(Customer* customer);
+        string getRating();
 };
 
 #endif
