+++ code/src/Unhappy.h
@@ -1,10 +1,12 @@
 /**
  * @file Unhappy.h
  * @brief Contains declaration for the Unhappy class
- * @authors Aidan Chapman (u22738917)
+ * @authors Aidan Chapman (u22738917), Sange Tshakumane (u21479748)
 */
 
 #include "Rating.h"
+
+using namespace std;
 
 #ifndef UNHAPPY_H
 #define UNHAPPY_H
@@ -15,6 +17,8 @@ class Unhappy : public Rating
         Unhappy();
         ~Unhappy();
         float calculateTip();
+        void changeState(Customer* customer);
+        string getRating();
 };
 
 #endif
